requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
aiortc>=1.6.0
numpy>=1.24.0
av>=10.0.0
Pillow>=10.0.0
websockets>=12.0
google-cloud-vision>=3.7.0
pylibsrtp>=0.8.0
google-genai>=0.8.0
python-dotenv>=1.0.0
google-cloud-storage>=2.10.0
google-cloud-firestore>=2.13.0
